refactor airflow
