+++ python_modules/libraries/dagster-airflow/dagster_airflow/__init__.py
@@ -9,11 +9,7 @@ from .dagster_job_factory import (
     make_dagster_definitions_from_airflow_dags_path as make_dagster_definitions_from_airflow_dags_path,
     make_dagster_definitions_from_airflow_example_dags as make_dagster_definitions_from_airflow_example_dags,
     make_dagster_job_from_airflow_dag as make_dagster_job_from_airflow_dag,
-)
-from .dagster_pipeline_factory import (
-    make_dagster_repo_from_airflow_dag_bag as make_dagster_repo_from_airflow_dag_bag,
-    make_dagster_repo_from_airflow_dags_path as make_dagster_repo_from_airflow_dags_path,
-    make_dagster_repo_from_airflow_example_dags as make_dagster_repo_from_airflow_example_dags,
+    make_schedules_and_jobs_from_airflow_dag_bag as make_schedules_and_jobs_from_airflow_dag_bag,
 )
 from .hooks.dagster_hook import DagsterHook as DagsterHook
 from .links.dagster_link import DagsterLink as DagsterLink
@@ -26,10 +22,9 @@ from .version import __version__ as __version__
 check_dagster_package_version("dagster-airflow", __version__)
 
 __all__ = [
-    "make_dagster_repo_from_airflow_dags_path",
     "make_dagster_definitions_from_airflow_dags_path",
     "make_dagster_definitions_from_airflow_dag_bag",
-    "make_dagster_repo_from_airflow_dag_bag",
+    "make_schedules_and_jobs_from_airflow_dag_bag",
     "make_dagster_job_from_airflow_dag",
     "load_assets_from_airflow_dag",
     "DagsterHook",
