+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_job_factory/test_load_dag_bag.py
@@ -156,10 +156,7 @@ def test_retry_conversion():
         dag_bag = DagBag(dag_folder=tmpdir_path)
         retry_dag = dag_bag.get_dag(dag_id="retry_dag")
 
-        job = make_dagster_job_from_airflow_dag(
-            dag=retry_dag,
-            use_ephemeral_airflow_db=True,
-        )
+        job = make_dagster_job_from_airflow_dag(dag=retry_dag)
         result = job.execute_in_process()
         assert result.success
         for event in result.all_events:
