+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_pipeline_factory/test_dependency_structure_translation.py
@@ -15,7 +15,6 @@ else:
 from dagster._core.snap import PipelineSnapshot
 from dagster._serdes import serialize_pp
 from dagster_airflow.dagster_job_factory import make_dagster_job_from_airflow_dag
-from dagster_airflow.dagster_pipeline_factory import make_dagster_pipeline_from_airflow_dag
 
 default_args = {
     "owner": "dagster",
@@ -44,7 +43,7 @@ def test_one_task_dag(snapshot):
     snapshot.assert_match(
         serialize_pp(
             PipelineSnapshot.from_pipeline_def(
-                make_dagster_pipeline_from_airflow_dag(dag=dag)
+                make_dagster_job_from_airflow_dag(dag=dag)
             ).dep_structure_snapshot
         )
     )
@@ -75,7 +74,7 @@ def test_two_task_dag_no_dep(snapshot):
     snapshot.assert_match(
         serialize_pp(
             PipelineSnapshot.from_pipeline_def(
-                make_dagster_pipeline_from_airflow_dag(dag=dag)
+                make_dagster_job_from_airflow_dag(dag=dag)
             ).dep_structure_snapshot
         )
     )
@@ -108,7 +107,7 @@ def test_two_task_dag_with_dep(snapshot):
     snapshot.assert_match(
         serialize_pp(
             PipelineSnapshot.from_pipeline_def(
-                make_dagster_pipeline_from_airflow_dag(dag=dag)
+                make_dagster_job_from_airflow_dag(dag=dag)
             ).dep_structure_snapshot
         )
     )
@@ -151,7 +150,7 @@ def test_diamond_task_dag(snapshot):
     snapshot.assert_match(
         serialize_pp(
             PipelineSnapshot.from_pipeline_def(
-                make_dagster_pipeline_from_airflow_dag(dag=dag)
+                make_dagster_job_from_airflow_dag(dag=dag)
             ).dep_structure_snapshot
         )
     )
@@ -194,7 +193,7 @@ def test_multi_root_dag(snapshot):
     snapshot.assert_match(
         serialize_pp(
             PipelineSnapshot.from_pipeline_def(
-                make_dagster_pipeline_from_airflow_dag(dag=dag)
+                make_dagster_job_from_airflow_dag(dag=dag)
             ).dep_structure_snapshot
         )
     )
@@ -236,7 +235,7 @@ def test_multi_leaf_dag(snapshot):
     snapshot.assert_match(
         serialize_pp(
             PipelineSnapshot.from_pipeline_def(
-                make_dagster_pipeline_from_airflow_dag(dag=dag)
+                make_dagster_job_from_airflow_dag(dag=dag)
             ).dep_structure_snapshot
         )
     )
@@ -492,7 +491,7 @@ def test_complex_dag(snapshot):
     snapshot.assert_match(
         serialize_pp(
             PipelineSnapshot.from_pipeline_def(
-                make_dagster_pipeline_from_airflow_dag(dag=dag)
+                make_dagster_job_from_airflow_dag(dag=dag)
             ).dep_structure_snapshot
         )
     )
