+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_pipeline_factory/test_load_dag_bag.py
@@ -3,9 +3,9 @@ import tempfile
 
 import pytest
 from airflow import __version__ as airflow_version
-from dagster_airflow.dagster_pipeline_factory import (
-    make_dagster_repo_from_airflow_dags_path,
-    make_dagster_repo_from_airflow_example_dags,
+from dagster_airflow.dagster_job_factory import (
+    make_dagster_definitions_from_airflow_dags_path,
+    make_dagster_definitions_from_airflow_example_dags,
 )
 
 from dagster_airflow_tests.marks import requires_airflow_db
@@ -30,17 +30,16 @@ def test_make_repo(
                 f.write(bytes(content.encode("utf-8")))
 
         repo = (
-            make_dagster_repo_from_airflow_dags_path(
-                tmpdir_path, repo_name, use_ephemeral_airflow_db=False
+            make_dagster_definitions_from_airflow_dags_path(
+                tmpdir_path,
             )
             if fn_arg_path is None
-            else make_dagster_repo_from_airflow_dags_path(
-                os.path.join(tmpdir_path, fn_arg_path), repo_name, use_ephemeral_airflow_db=False
+            else make_dagster_definitions_from_airflow_dags_path(
+                os.path.join(tmpdir_path, fn_arg_path),
             )
-        )
+        ).get_repository_def()
 
         for job_name in expected_job_names:
-            assert repo.name == repo_name
             assert repo.has_job(job_name)
 
             job = repo.get_job(job_name)
@@ -108,10 +107,9 @@ def test_airflow_example_dags(
     expected_job_names,
     exclude_from_execution_tests,
 ):
-    repo = make_dagster_repo_from_airflow_example_dags()
+    repo = make_dagster_definitions_from_airflow_example_dags().get_repository_def()
 
     for job_name in expected_job_names:
-        assert repo.name == "airflow_example_dags_repo"
         assert repo.has_job(job_name)
 
         if job_name not in exclude_from_execution_tests:
