+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_pipeline_factory/test_solid_execution.py
@@ -26,7 +26,7 @@ from dagster._core.storage.compute_log_manager import ComputeIOType
 from dagster._core.test_utils import instance_for_test
 from dagster._legacy import execute_pipeline
 from dagster._seven import get_current_datetime_in_utc
-from dagster_airflow.dagster_pipeline_factory import make_dagster_pipeline_from_airflow_dag
+from dagster_airflow.dagster_job_factory import make_dagster_job_from_airflow_dag
 
 default_args = {
     "owner": "dagster",
@@ -55,16 +55,16 @@ def test_normalize_name():
         dag=dag,
     )
 
-    pipeline_def = make_dagster_pipeline_from_airflow_dag(
+    job_def = make_dagster_job_from_airflow_dag(
         dag=dag,
         tags={AIRFLOW_EXECUTION_DATE_STR: get_current_datetime_in_utc().isoformat()},
     )
-    result = execute_pipeline(pipeline_def)
+    result = job_def.execute_in_process()
 
     assert result.success
-    assert result.pipeline_def.name == "airflow_dag_with_dot_dash"
-    assert len(result.pipeline_def.solids) == 1
-    assert result.pipeline_def.solids[0].name == "airflow_task_with_dot_dash"
+    assert result.job_def.name == "airflow_dag_with_dot_dash"
+    assert len(result.job_def.solids) == 1
+    assert result.job_def.solids[0].name == "airflow_task_with_dot_dash"
 
 
 # Test names with 250 characters, Airflow's max allowed length
@@ -88,21 +88,21 @@ def test_long_name():
         dag=dag,
     )
 
-    pipeline_def = make_dagster_pipeline_from_airflow_dag(
+    job_def = make_dagster_job_from_airflow_dag(
         dag=dag,
         tags={AIRFLOW_EXECUTION_DATE_STR: get_current_datetime_in_utc().isoformat()},
     )
-    result = execute_pipeline(pipeline_def)
+    result = job_def.execute_in_process()
 
     assert result.success
     assert (
-        result.pipeline_def.name
+        result.job_def.name
         == "airflow_dag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ongdag_with_dot_dash_lo00ong"
     )
 
-    assert len(result.pipeline_def.solids) == 1
+    assert len(result.job_def.solids) == 1
     assert (
-        result.pipeline_def.solids[0].name
+        result.job_def.solids[0].name
         == "airflow_task_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loongtask_with_dot_dash2_loong"
     )
 
@@ -125,11 +125,11 @@ def test_one_task_dag():
         dag=dag,
     )
 
-    pipeline_def = make_dagster_pipeline_from_airflow_dag(
+    job_def = make_dagster_job_from_airflow_dag(
         dag=dag,
         tags={AIRFLOW_EXECUTION_DATE_STR: get_current_datetime_in_utc().isoformat()},
     )
-    result = execute_pipeline(pipeline_def)
+    result = job_def.execute_in_process()
     assert result.success
 
 
@@ -189,19 +189,18 @@ def test_template_task_dag():
         execution_date_add_one_week = execution_date + datetime.timedelta(days=7)
         execution_date_iso = execution_date.isoformat()
 
-        result = execute_pipeline(
-            make_dagster_pipeline_from_airflow_dag(
-                dag=dag, tags={AIRFLOW_EXECUTION_DATE_STR: execution_date_iso}
-            ),
-            instance=instance,
+        job_def = make_dagster_job_from_airflow_dag(
+            dag=dag, tags={AIRFLOW_EXECUTION_DATE_STR: execution_date_iso}
         )
+        result = job_def.execute_in_process(instance=instance)
+
         assert result.success
-        for event in result.step_event_list:
+        for event in result.all_events:
             assert event.event_type_value != "STEP_FAILURE"
 
         capture_events = [
             event
-            for event in result.event_list
+            for event in result._event_list
             if event.event_type == DagsterEventType.LOGS_CAPTURED
         ]
         assert len(capture_events) == 1
@@ -287,12 +286,11 @@ def test_spark_dag(mock_subproc_popen):
         conn_id="SPARK",
         dag=dag,
     )
-
-    pipeline = make_dagster_pipeline_from_airflow_dag(
+    job_def = make_dagster_job_from_airflow_dag(
         dag=dag,
         tags={AIRFLOW_EXECUTION_DATE_STR: get_current_datetime_in_utc().isoformat()},
     )
-    execute_pipeline(pipeline)
+    job_def.execute_in_process()
 
     if airflow_version >= "2.0.0":
         assert mock_subproc_popen.call_args_list[0][0] == (
